1) Refactored to move write_delimited_pb call from all methods to a single ::write() call. 2) ::write() method just calls F(message*, ZeroCopyStream*) 3) bool (*F)(google::protobuf::MessageLite*, google::protobuf::ZeroCopySteam*) is a template argument. 4) other writers (e.g.-jsonlines) can be implemented by passing a different F function. 5) split the tests to properly either do 1) a OstreamOutputStream to a temporary stringstream; or 2) a FileOutputStream to a file descriptor. Both are faster.  The OstreamOutputStream example is still contrived since but it's a test...
