+++ Makefile
@@ -3,7 +3,7 @@ default: test
 
 libraries: lib/libstanc.a lib/libgtest.a lib/libprotobuf.a
 
-test-binaries: test/unit/compile-test test/unit/binary_proto_stream_writer-test test/unit/rw_delimited_pb-test
+test-binaries: test/unit/compile-test test/unit/rw_delimited_pb-test test/unit/binary_proto_filestream_writer-test test/unit/binary_proto_ostream_writer-test
 
 generated:
 	./lib/protobuf/src/protoc --cpp_out=src/stan/proto --proto_path=src/stan/proto src/stan/proto/*.proto
@@ -12,7 +12,8 @@ test: libraries generated test-binaries
 	@echo running tests
 	test/unit/compile-test
 	test/unit/rw_delimited_pb-test
-	test/unit/binary_proto_stream_writer-test
+	test/unit/binary_proto_filestream_writer-test
+	test/unit/binary_proto_ostream_writer-test
 
 test-full: test cpplint
 
@@ -48,7 +49,7 @@ test/unit/rw_delimited_pb-test: src/test/rw_delimited_pb-test.cpp
 			lib/libstanc.a \
 			lib/libprotobuf.a
 
-test/unit/binary_proto_stream_writer-test: src/test/binary_proto_stream_writer-test.cpp
+test/unit/binary_proto_filestream_writer-test: src/test/binary_proto_filestream_writer-test.cpp
 	mkdir -p test/unit
 	g++ -std=c++0x -I lib/stan/lib/stan_math/lib/gtest_1.7.0/include \
 			-I src \
@@ -56,8 +57,25 @@ test/unit/binary_proto_stream_writer-test: src/test/binary_proto_stream_writer-t
 			-I lib/protobuf/src \
 			-isystem lib/stan/lib/stan_math/lib/boost_1.58.0 \
 			-pthread \
-			-o test/unit/binary_proto_stream_writer-test \
-			src/test/binary_proto_stream_writer-test.cpp \
+			-o test/unit/binary_proto_filestream_writer-test \
+			src/test/binary_proto_filestream_writer-test.cpp \
+			src/stan/proto/stan-core.pb.cc \
+			src/stan/proto/compile.pb.cc \
+			src/stan/proto/sample.pb.cc \
+			lib/libgtest.a \
+			lib/libstanc.a \
+			lib/libprotobuf.a
+
+test/unit/binary_proto_ostream_writer-test: src/test/binary_proto_ostream_writer-test.cpp
+	mkdir -p test/unit
+	g++ -std=c++0x -I lib/stan/lib/stan_math/lib/gtest_1.7.0/include \
+			-I src \
+			-I lib/stan/src \
+			-I lib/protobuf/src \
+			-isystem lib/stan/lib/stan_math/lib/boost_1.58.0 \
+			-pthread \
+			-o test/unit/binary_proto_ostream_writer-test \
+			src/test/binary_proto_ostream_writer-test.cpp \
 			src/stan/proto/stan-core.pb.cc \
 			src/stan/proto/compile.pb.cc \
 			src/stan/proto/sample.pb.cc \
